requirements.txt
rel
websocket-client
asyncio
numpy
gevent
eventlet